requirements.txt
blinker==1.8.2
boto3==1.35.21
botocore==1.35.21
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
Flask==3.0.3
Flask-Cors==4.0.1
Flask-Mail==0.10.0
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
jmespath==1.0.1
MarkupSafe==2.1.5
mysql-connector-python==9.0.0
pdfkit==1.0.0
pillow==11.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
qrcode==8.0
requests==2.32.3
s3transfer==0.10.2
sentry-sdk==2.13.0
six==1.16.0
urllib3==2.2.2
Werkzeug==3.0.3
